@@ -1,10 +1,9 @@
-import Vue from "vue";
-import App from "./App.vue";
-import store from "./store";
+import { createApp } from 'vue'
+import App from './App.vue'
+import store from './store'
 
-Vue.config.productionTip = false;
+console.log('from main:', store.state)
 
-new Vue({
-  store,
-  render: h => h(App)
-}).$mount("#app");
+createApp(App)
+  .use(store)
+  .mount('#app')
